ci: start testing Node 18 future-proofed using `lts` keyword
BREAKING CHANGE: no longer supports Node 12

Various libraries are starting to drop support for Node 12.
While this only affects the dev dependencies of this project
it's only a matter of time before something critical drops
support. So to keep maintenance easy, we're dropping
v12 now as it is past EOL already and moving to 14, 16,
and 18 LTS versions.

Recommended action: upgrade your project from Node 12
to a later LTS release (14, 16, 18)